+++ types/configs/methods.go
@@ -4,7 +4,7 @@ import "github.com/statping/statping/utils"
 
 // Save will initially create the config.yml file
 func (d *DbConfig) Save(directory string) error {
-	if err := utils.Params.SafeWriteConfigAs(directory + "/config.yml"); err != nil {
+	if err := utils.Params.WriteConfigAs(directory + "/config.yml"); err != nil {
 		return nil
 	}
 	return nil