+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.90.30 (04-19-2020)
 - Attempt to fix Github Actions build process
+- Fix for empty database connection string, and not starting in setup mode
 
 # 0.90.29 (04-19-2020)
 - Added HTTP Redirects for services